worker/requirements.txt
Pillow==10.2.0
marker-pdf==1.0.2
pix2tex==0.1.3
ollama==0.4.4
pika==1.3.2
psycopg[binary]==3.2.3
